requirements.txt
Flask==3.1.3
gunicorn==22.0.0
python-dotenv==1.0.1
